Organise Wasm/JS interface code
