+++ src/wasm-js/generated.d.ts
@@ -1,0 +1,16 @@
+/**
+ * App-specific values generated by the Elm compiler
+ * to facilitate JS/Wasm communication
+ */
+interface NameToAddress {
+  [name: string]: number;
+}
+interface AddressToName {
+  [addr: number]: string;
+}
+
+export declare const appTypes: {
+  fields: NameToAddress & AddressToName;
+  fieldGroups: NameToAddress & AddressToName;
+  ctors: NameToAddress & AddressToName;
+};
